# swagger.yaml
openapi: 3.0.0
info:
  title: Saint-Pierre API
  version: 1.0.0
  description: API complète pour la gestion des devoirs, cours et messagerie scolaire
  contact:
    name: Support
  x-logo:
    url: https://your-api-logo.png
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.saint-pierre.com
    description: Serveur de production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/sign-in
  schemas:
...
    Rendu:
      type: object
      properties:
        id_rendu:
          type: string
          pattern: '^[0-9]+$'
          example: '1'
        id_devoir:
          type: string
          pattern: '^[0-9]+$'
        id_user:
          type: string
          pattern: '^[0-9]+$'
        date_rendu:
          type: string
          format: date-time
          nullable: true
        note:
          type: number
          format: decimal
          nullable: true
          example: 18.5
        retour:
          type: string
          maxLength: 255
          nullable: true
        archive:
          type: boolean
          default: false
        pieces_jointes:
          type: array
          items:
            $ref: '#/components/schemas/PieceJointeRendu'
    PieceJointeRendu:
      type: object
      properties:
        id_piece_jointe:
          type: string
          pattern: '^[0-9]+$'
        nom_fichier:
          type: string
          maxLength: 255
        chemin_fichier:
          type: string
        type_fichier:
          type: string
        taille_octets:
          type: string
          pattern: '^[0-9]+$'
        date_ajout:
          type: string
          format: date-time

    # ========== ÉVÉNEMENT ==========
...
    Error400:
      type: object
      properties:
        error:
          type: string
          example: 'Paramètres invalides'
...
    Error404:
      type: object
      properties:
        error:
          type: string
          example: 'Ressource non trouvée'
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Opération réussie'
...
security:
  - BearerAuth: []
paths:
...
  /api/rendus:
    post:
      summary: Rendre un devoir (avec fichiers)
      tags: [Rendus]
      operationId: submitDevoir
      x-roles: [eleve]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id_devoir:
                  type: string
                  pattern: '^[0-9]+$'
                id_user:
                  type: string
                  pattern: '^[0-9]+$'
                fichiers:
                  type: array
                  items:
                    type: string
                    format: binary
              required: [id_devoir, id_user]
      responses:
        '201':
          description: Devoir rendu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rendu'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
  /api/rendus/devoir/{idDevoir}:
    delete:
      summary: Supprimer le rendu d'un devoir
      tags: [Rendus]
      operationId: deleteRendu
      parameters:
        - name: idDevoir
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Rendu supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Rendu non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
  /api/rendus/{idRendu}:
    put:
      summary: Mettre à jour le rendu (note et retour - Professeur)
      tags: [Rendus]
      operationId: updateRendu
      x-roles: [professeur, administrateur]
      parameters:
        - name: idRendu
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: number
                  format: decimal
                retour:
                  type: string
      responses:
        '200':
          description: Rendu mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rendu'
        '404':
          description: Rendu non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
...
  /api/rendus/{idRendu}/archive:
    post:
      summary: Archiver un rendu
      tags: [Rendus]
      operationId: archiveRendu
      parameters:
        - name: idRendu
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Rendu archivé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rendu'
        '404':
          description: Rendu non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'

  # ========== ÉVÉNEMENTS ==========
...